[sdk/python] Add an `invoke_async` function (#15602)
This commit adds a new `pulumi.runtime.invoke_async` function that
allows calling invokes asynchronously. The initial intended use for this
is inside the Kubernetes Python SDK's `yaml.ConfigFile`,
`yaml.ConfigGroup`, `helm`, `kustomize`, etc. components to avoid stalls
in resource registrations which severely limits parallelism.

I'd love to add some kind of benchmark, perhaps along the lines of the
repro in #15462, but we can add that as a fast-follow.

Part of #15462 and #15591

After this is merged and released, we can ship an updated version of the
Kubernetes Python SDK that makes use of it.