clean up sdks dir for each conformance test (#15604)
Multiple conformance tests can use the same provider. When a provider
gets used for a test we install all its dependencies. Then in subsequent
runs, when the provider is re-used those dependencies are already in the
directory, and we complain that those newly installed dependencies don't
exist in the original snapshot.

Fix this by cleaning up the sdks directory after each run, so we get a
clean slate each time we codegen a provider.

We could potentially consider getting a list of providers when preparing
the language tests, and just generating their code once, but that's a
little bit more involved.

Fixes https://github.com/pulumi/pulumi/issues/15603