Support suppress-progress and suppress-outputs options (#15596)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->
Introduces support for `--suppress-progress` and `--suppress-outputs`
parameters of the cli for stack `up`, `destroy`, `preview`, and
`refresh`.

Fixes https://github.com/pulumi/pulumi/issues/12549
precondition for: https://github.com/pulumi/actions/issues/1108

## Checklist
I'm not running any of the ones below as it's at least shady what it
requires and the way it runs

- [x] I have run `make tidy` to update any new dependencies
- [ ] I have run `make lint` to verify my code passes the lint check _->
fails with `ERRO Running error: unknown linters: 'perfsprint', run
'golangci-lint help linters' to see the list of supported linters `_
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works _-> current params dont seem to be tested in any way, and
it's not easily testable as its directly dependent on the cli_
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>