tag pkg automatically again
We disabled this because the way we were tagging the version was
incorrect as we were tagging a version outside of the main branch, but
the way we fixed it the first time around (trying to push the commits
directly to master) didn't work because of branch protections.

Fix this now by tagging pkg during merge, using the version we get
from the commit message.