Await output properties before closing the context (#15611)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

This PR fixes a race condition between resolving the output properties
(from `Construct` or `Call`) and the closing of the context. The
solution is to await the output properties before closing the context.

New test cases were added for `Convert` and `Call`, with
`TestConstruct_await/unassociated_output` and
`TestCall_await/unassociated_output` being the specific repro cases.
Without the fix, the test fails as expected:
```
=== RUN   TestConstruct_await/unassociated_output
    /Users/eronwright/Pulumi/pulumi/sdk/go/pulumi/provider_test.go:2027:
                Error Trace:    /Users/eronwright/Pulumi/pulumi/sdk/go/pulumi/provider_test.go:2027
                Error:          Received unexpected error:
                                marshaling properties: awaiting input property "b": attempted illegal RPC after program completion
                Test:           TestConstruct_await/unassociated_output
                Messages:       failed to construct
--- FAIL: TestConstruct_await/unassociated_output (1.00s)
```

Fixes #15527

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->