tag pkg automatically again (#15629)
We disabled this because the way we were tagging the version was
incorrect as we were tagging a version outside of the main branch, but
the way we fixed it the first time around (trying to push the commits
directly to master) didn't work because of branch protections.

Fix this now by tagging pkg during merge, using the version we get from
the commit message.

I think we can actually test this one by creating a fake "Changelog and
go.mod updates for v0.0.9999" PR, merging that, and deleting that tag
again afterwards. It's going to be after the fact testing, but at least
we should be able to do it before the release.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>