mark snapshot tests as non-parallel (#15642)
These tests are currently flaky when run in parallel. Since the proper
fix is a bit more involved, and we really want to turn of rerunning
tests on failures, make them not run in parallel for the time being.
We'll keep https://github.com/pulumi/pulumi/issues/15461 open to remind
us to fix this properly and get them running in parallel again.