do dev releases post merge
Currently we do our dev releases while running CI for the merge queue,
which adds to the time it takes for things to get through the merge
queue, slowing our development process down.

However there's no really good reason to do dev releases pre-merge.
The pre-merge checks are there to make sure the main branch is always
green.  We already know the binaries will be able to be built
successfully since we do that for integration tests.  And if the
network fails for uploading the binaries we can always just retry the
CI job, which we can do for a on-push CI job, but isn't useful for the
merge CI job since the PR will already have been taken out of the
queue.