destroy: implement --continue-on-error
Implement a `--continue-on-error` flag for `pulumi destroy`.  This
makes sure we continue processing destroys even if errors occur.  We
will only continue for resources which do not depend on the failed
resource, to make sure we always have a valid snapshot available, and
to not leave any orphaned resources behind.

Resources that fail to destroy will still continue to be managed by
pulumi, and the process ends in an error to indicate to the user that
there were problems and the destroy didn't succeed cleanly.

The output will continue to be the same as for failed destroys, except
we now may succeed in destroying more resources than before.