Asset and Archive can have missing contents (#15736)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Fixes https://github.com/pulumi/pulumi/issues/15729.

This fixes Assets and Archives to allow four states, as opposed to the
three that https://github.com/pulumi/pulumi/pull/14007 had enforced.

An asset can either be Text, Path, Uri, or none. That is `IsText`,
`IsPath`, and `IsURI` can all return false. Similarly for archives
except with Assets instead of Text.
This happens when a provider returns an Assert (or Archive) with a hash
value set, but no other contents.

The Asset and Archive objects have been updated to handle this case, and
a number of places in the CLI that asserted that one of
IsText/IsPath/IsURI were true have been fixed up to handle the case of
all three being false.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->