Make TestConstructComponentConfigureProviderPython less flaky (#15760)
Update TLS library to a more recent version that doesn't have the old
"install the plugin as part of setup.py" logic.

Then enable the engine to do the plugin download for this test by
disabling `PULUMI_DISABLE_AUTOMATIC_PLUGIN_ACQUISITION`. Should make the
test less flaky as the engine is managing the download and can retry,
but at least makes the error much more clear if it does fail.