Make local dependencies absolute or relative paths
We need this for local dependencies in conformance tests for SDKs and programs.
Things like NodeJS just record the path as is in the package.json put into the
packed tgz, if a program then uses that tgz via a relative path NodeJS doesn't
re-resolve the relative relative paths.
Making these absolute for conformance testing fixes that as all the conformance
tests do run in a stable folder location. We fix writing that folder path to
the snapshots using the new regex replace facility added in
https://github.com/pulumi/pulumi/pull/15747.